Fix some compiler warnings courtesy of OSX.

+++ src/raw.c
@@ -61,7 +61,7 @@ off64_t _GD_RawSeek(struct _gd_raw_file* file, off64_t count,
 ssize_t _GD_RawRead(struct _gd_raw_file *file, void *ptr, gd_type_t data_type,
     size_t nmemb)
 {
-  int nread;
+  size_t nread;
 
   dtrace("%p, %p, %x, %zu", file, ptr, data_type, nmemb);
